Add todo lists?
Okay so obviously I have no idea what I'm doing and I don't even know
Objective-C. However, I wanted to try and start digging into it and
created this to mimick Automattic/simplenote-electron#add/todo-lists

I'm lost and would love some feedback. In Simplenote-electron I kept
track of how many times each match was found from the RegExp and then
used that index as a reference back into the app to update the note to
reflect clicks.

Here in iOS I just have no idea. The main idea is to reference back out
of the WebView so we can update the note but I'm not sure how to get to
the note in the `SPMarkdownPreviewViewController` let alone update it.

This will probably die a painful death but thanks for looking!