Define lane to take screenshots
Notice how I also removed the Snapfile. This is in line with what the
WooCommerce iOS does.

It seems a great way to keep all the configuration in a single place,
and have a clear and unique source of truth on how to generate
screenshots. In fact, if we had a lane and a Snapfile, it might be
unclear to know whether to add a setting to the lane or the file, or
whether to take screenshots via `snapshot` or the lane itself.

Finally, the output folder for the screenshots is now nested in
`fastlane/` so it's consistent with all the other folders in this
Fastfile setup.