Add promo screenshot mask for iPhone Xs Max
Applying the mask showed a bit of asymmetry in the screenshot /
device frame overlap. Looking into it, it seemed to me like the frame
hasn't got the exact dimensions, so I updated its size in the
screenshot to compensate for that (stretch it on the y axis).

Also notice how the height of the screenshot itself went from 1927 to
1926. I don't remember how I came up with the former value, probably
some math plus manual adjustment. 1926 is more correct, though:

  890 / 1242 = x / 2688

  890 * 2688 / 1242 = x

  x = 1926.183574879