Update Fastlane iOS Simulator version for screenshots to 13.6
This is due to us using macos-latest in the GitHub Actions workflow. It
might be better to switch to a targeted macOS image instead.

Also notice we had to update the snapshot helper file. That was prompted
by a build failure which I think is related to the new Simulator version
only, since we didn't update Fastlane since the last screenshots build.

See https://github.com/Automattic/simplenote-ios/runs/815164553#step:9:1103