Update RELEASE-NOTES.txt
When #824 was merged, version 4.22 had already been put in code freeze,
but the change had not been merged into develop yet. That resulted in
not entry for version 4.23 being available in the release notes file,
so the change from that PR ended up as part of 4.22.

This rectifies the issue.