Add note regarding tap on pin screen in screenshots automation
I've been seeing inconsistent failures in this part of the test and have
experimented with how the interaction is done to try and fix them.
Unfortunately, I don't think my fiddling made any difference.

This note regarding the UI structure and how it reflects in the matches
to the query is useful, though.