Remove lane to trim keywords text files longer than 100 chars
Trimming too long keywords client side is the wrong solution to the
problem of having keywords that are too long.

Because keywords change sporadically, it's much better to edit the
localization source on the GlotPress end.

And even better long term fix would be to add a character limit to
GlotPress. I don't know how to do that know but I'll get in touch with
folks that might help with that.

See discussion at
https://github.com/Automattic/simplenote-ios/pull/1201#discussion_r590997663