Create new framework to host UITestHelpers
I decided not to add a test target for the moment.

When we'll extract this into a dedicated project, we shall add at least
a dummy UI app to test it.

For the moment, since it will only contains code consumed by the test
targets in this project, we can skip the tests.