Replace UI tests helper local framework with remote Swift package
As part of the 2nd HACK week for 2021, I'm working on UI testing shared
infrastructure.

As step towards sharing more code across projects was extracting
general-purpose helpers present in the Simplenote iOS project into a
dedicated framework.

The second was to create a [Swift
package](https://github.com/Automattic/SwiftUITestsHelpers) from that
code.

With this diff, we replace the local framework with its Swift package
version.