Add ScreenObject Swift package – Link to screenshots target
I decided to link it only to the screenshots automation target because
the work required to meaningfully integrate the framework in the UI
tests target is more complex.

At this point in time, my goal is simply to get the framework setup and
used in a leaf screen within the navigation tree, just to serve as a
proof of concept.