Remove empty and unused InfoPlist.strings file
The file reference being moved under the new "Supporting Files" group in Xcode made the compiler fail to find it
(looking for it in `Supporting Files/en.lproj/InfoPlist.strings` since it's a localized resource) during build/archive.

I was about to move the reference back to "relative to project", but saw that the file was actually empty and we don't translate any Info.plist metadata/key… which means we can just get rid of it in the first place.