Remove `previous-application-identifiers` node from widgets entitlements
I landed on this change by trial and error.

I looked around the Developer Portal and `match` but could not find a
way to add the entitlement to the provisioning profile, so I tried to
remove the value from the entitlements file and see what happened. The
result was that the app archived successfully and passed the App Store
connect validation that one can initiate from the Xcode organizer.

Apart from the fact that, so far, this configuration seems to be working
fine, I think it is appropriate because of the
`previous-application-identifiers` entitlement purpose.

The name `previous-application-identifiers` clearly suggests what the
value is for, but I did some research to verify it and landed on this
documentation page:
https://developer.apple.com/library/archive/technotes/tn2319/_index.html#//apple_ref/doc/uid/DTS40013778-CH1-ERRORMESSAGES-UPGRADE_S_APPLICATION_IDENTIFIER_DOES_NOT_MATCH_THE_INSTALLED_APP

The page is archived, so possibly out of date, confirms the natural
guess that the value is to be used for when changing the App ID Prefix
for an application.

Simplenote is an Automattic acquisition, so it's not surprising we might
have needed to include that entitlement at some point.

The reason I don't think it's required for the widgets extension,
though, is that application ID was created after the migration to the
Automattic organization in App Store Connect. The extension is only a
few months old, see 8ac38788.

The documentation page I linked above might also explain why I couldn't
find a way to add the entitlement to the provisioning profile. It seems
to be an ad hoc manoeuvre that only Apple can perform:

> Request a special provisioning profile from Developer Programs that
will also contain the above entitlement. Make this request through the
Developer Contact page > Membership > Enrollment and Account > Apple
Developer Program Support form.

Some of the answers to this related StackOverflow question also report
the need to get in touch with Apple:
https://stackoverflow.com/questions/29478359/ios-application-update-failed-after-app-transfer-with-rejecting-upgrade-error

The only flaw in this theory of mine is that the provisioning profile
for the intents extension _does include_ that entitlement! I don't have
  an explanation for that discrepancy...

I think we should proceed with this solution on the grounds that neither
Xcode nor App Store Connect are showing errors and see if it works.

Researching odd edge case in the configuration and interaction with
Apple's backend when it comes to code signing is interesting but also
has diminishing returns.